append GPL

+++ src/components/public/HeaderNav.vue
@@ -10,7 +10,10 @@
         <a href="javascript:;" @click="exitLogin">退出</a>
       </div>
       <div class="right">
-        <router-link to="/cart/cartlist">购物车</router-link>
+        <router-link to="/cart/cartlist">
+          <span class="iconfont icon">&#xe600;</span>
+          购物车{{userName?'('+count+')':'('+0+')'}}
+        </router-link>
         <router-link to="/myMall/myOrder">我的订单</router-link>
         <router-link to="/myMall/userCenter">我的Mall</router-link>
         <router-link to="/aboutMall">关于Mall</router-link>
@@ -24,25 +27,61 @@ export default {
   name: "headerNav",
   data() {
     return {
+      count: "",
       userName: "",
-      isLogin: ""
+      isLogin: "",
     };
   },
   methods: {
     exitLogin() {
-      this.userName =  localStorage.clear("userName");
-      this.isLogin =  localStorage.clear("isLogin");
+      this.$axios
+        .post("/users/logout")
+        .then(result => {
+          let code = result.data.code;
+          if (code === 1) {
+            this.userName = localStorage.clear("userName");
+            this.isLogin = localStorage.clear("isLogin");
+          }
+        })
+        .catch(err => {});
+    },
+    getCartData() {
+      var userName = localStorage.getItem("userName");
+      if(!userName){
+        return
+      }
+      this.$axios
+        .get("/cart/cartData", {
+          params: {
+            userName: userName
+          }
+        })
+        .then(result => {
+          var code = result.data.code;
+          if (code === 1) {
+            this.count = result.data.result.length;
+          } else {
+            alert("获取购物车数据失败");
+          }
+        })
+        .catch(err => {
+          alert("获取购物车数据失败");
+        });
       
     }
   },
   created() {
     this.userName = localStorage.getItem("userName");
     this.isLogin = localStorage.getItem("isLogin");
+    this.getCartData();
   }
 };
 </script>
 
 <style lang="css" scoped>
+.icon {
+  margin: 2px;
+}
 .nav {
   width: 100%;
   background-color: #eee;
